// typings/contracts/PookyGame.ts
export type BallUpdatesStruct = {
  ballId: PromiseOrValue<BigNumberish>;
  addPxp: PromiseOrValue<BigNumberish>;
  toLevelUp: PromiseOrValue<boolean>;
};